requirements.txt
marshmallow==4.2.2
packaging==26.0
paho-mqtt==1.6.1
